Snap opportunistically in Modal Bottom Sheet
We now use the trySnapTo API to attempt to snap immediately, instead
of launching a coroutine that would most likely only run in the next
frame.
Since we rely on the anchor change handler to initialize the
Swipeable state by snapping if no anchor is found for the initial
value, we need to update the offset immediately, not only a frame
later.

Relnote: "Fixed an issue where ModalBottomSheetLayout would crash
in an edge case on orientation change. Layout animations (e.g.
`Modifier.animateContentSize`) in/on the sheet content now work
smoothly."
Test: modalBottomSheet_anchorChangeHandler_missingAnchor
_immediatelySnapsForInitialization
Bug: 266780234

Change-Id: I2f981d892f163254f0282d6874fb860aae5d0992