[RangeSlider] API Council Feedback
Introduce RangeSliderState which holds all of the information for the current active track for RangeSlider as well as all of the measurements of the components of the RangeSlider. RangeSliderState also holds the gesture and dragging logic.
Updating SliderDefaults.Track to use the RangeSliderState to calculate the values needed to draw the track.

Bug: 270969255
Test: Adding a RangeSliderRecompositionCounter that takes in a RangeSliderState, and updating the recomposition tests appropriately with this recomposition counter and state.
Relnote: Added an overloads for RangeSlider that are stateful version of the component. Created RangeSliderState that holds all of the information for the current active track, the measurements of the components of the RangeSlider, and the drag and gesture logic.
Change-Id: I8c270fd01856f713bb5a40bf356f2875e64fd6e1