Fix unboxValueIfInline for non-jvm targets
k/js and k/native can't unbox value class instances when the value class is declared in a different compilation unit and its property is private and/or its constructor is private.

If such a value class is in the same compilation unit then its instance should be unboxed.

The fix is to NOT unbox if it's not possible.

This CL also fixes the default arguments of value class types.
We can't create a "fake" default argument if the value class constructor is private.

Test: `./gradlew :compose:compiler:compiler-hosted:integration-tests:testDebugUnitTest` + tested k/js and k/native in compose-jb fork
Change-Id: I5d4de9f11ef0abd4ce4d5f76ab18c8fae3fdabb1