Fix NPE in AndroidView when in a SubcomposeLayout
This fixes an issue in G3 where AndroidView can occasionally crash with
an NPE when placed inside of a SubcomposeLayout (or Lazy Layout) when
using modifiers that change the size of the View. This only happens when
the AndroidView is an indirect child inside of the SubcomposeLayout. We
attempted to fix this previously with aosp/2464334, but we've now found
out how to reproduce the original issue in isolation.

The main fix here is to use `removeAllViewsInLayout()` which doesn't
trigger the AndroidViewHolder to request a layout. We're also adding
some other safeguards in place in case the system decides to measure
the View at any other time on its own.

Bug: b/271156218
Test: Added regression test to SubcomposeLayoutTest
Change-Id: Id7c290abe295f158336d6ee3d6772576bd7481c2