Enforce order of lookahead remeasurement vs. main pass remeasurement
Previously remeasure/relayout requests are not differentiated in
terms of whether they affect lookahead. As a result, when (non-lookahead)
remeasurement starts from higher up in the tree, it could potentially
get on the forceRemeasureSubtree code path and ignore new lookahead
requests due to its layoutState.

This CL ensures that lookahead requests are handled first. As such,
remeasure/relayout requests are now handled based on not only the
depth of the nodes, but also whether they affect lookahead.

Test: new test added.

Change-Id: Ie90e6eb571660f42941b59f99ee93563c946dee1