Fix Regression in custom enter/exit focus properties
We use a fetchCustomEnter function that has some logic to prevent the
lambda from being re-entrant when focus is redirected to a portion of
the hierarchy that already executed the enter/exit block.

When a return statement is placed in an anonymous lambda, it
short-circuits to return from the entire function, not from the lambda.
In this case it skipped over some cleanup logic
isProcessingCustomEnter = false inside the intermediate inline function,
so this led to a nonobvious bug.

Added a try-finally block to fix this issue and ensure that the cleanup
logic is always run.

Fixes: 277234245
Test: ./gradlew compose:ui:ui:cC -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.RequestFocusEnterTest
Test: ./gradlew compose:ui:ui:cC -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.RequestFocusExitTest
Change-Id: I7715190081686b2b01010b35375633cfc1c6ca3b