Hook animateContentSize & AnimatedVisibility into lookahead system
AnimatedVisibility and animateContentSize previously kept the
layout animation target to itself. In this CL, they report
their target size to the lookahead system. This therefore
allows lookahead system to see how the post-animation layouts
(from animateContentSize and AnimatedVisibility) would influence
the layout of the rest of the tree.

RelNote: "New property in IntrinsincMeasureScope and its implementations
 (e.g. MeasureScope) to indicate whether the current measure pass is
a lookahead pass."

Test: included new tests

Change-Id: I7a8122bf09752d7afb6ace1a0b397a2632708e95