Fixes AlertDialog dismiss actions with long text
This CL fix M2 and M3 AlertDialogs to ensure that when the dismiss and
confirm actions are stacked above each other in cases where their text
is long, the dismiss action appears on the bottom as specified by the
Material Design spec.
https://m3.material.io/components/dialogs/guidelines

Fixes: 235454277
Test: AlertDialogTest
Relnote: "Fixed the AlertDialog dismiss action to appear below the
confirm action when the actions stacked over each other to fit into the
dialog's width. This fix align the implementation with the Material
Design spec."

Change-Id: I029ded5c6dd79f38b1a060afb3d24dcfb9cf119a