Cache keys for lazy layouts
Currently it is possible that the key lambda passed to LazyColumn or other lazy layouts will be invoked multiple types for each item. It is not ideal for performance as sometimes the implementations of such lambdas are allocating new non trivial objects. Now we will be caching the keys near the currently visible range and only re-execute the lambdas when something changes.

Test: new tests for each lazy layout
Change-Id: I54ad3157ae41fbc6e19d8a840aec4bfcb9fa089a